Fix UI for Showtimes page
