+++ public/index.html
@@ -30,9 +30,11 @@
     <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/owl.carousel.min.css" />
     <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/style.css" />
     <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/css/responsive.css" />
-    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/revolution/css/colorbox.css" />
+
+    <!--Revolution css-->
     <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/revolution/css/custom-schedual.css" />
-    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/revolution/css/schedualPage.css" />
+    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/revolution/css/colorbox.css" />
+  
 <!-- ===== manifest.json links ===== -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
   </head>
