+++ src/components/MovieShowtimes/infoFilm.js
@@ -1,0 +1,70 @@
+import React, { Component } from 'react'
+import './../../styles/schedualPage.css'
+
+export default class extends Component {
+  render() {
+    return (
+
+
+      <div className="mainMaxWidth2 contentMain detailMain ">
+        <div id="homeNewsMain mt100">
+          <div className="row isFlex detailMainStyle ">
+            <div className="col-sm-6 col-xs-12 film left mt100 ">
+
+              {/* Start row Info */}
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Ngày phát hành</p>
+                <p className="contentInfo ng-binding">22.03.2019</p>
+              </div>
+              {/* Finish row Info */}
+              {/* Start row Info */}
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Đạo diễn</p>
+                <p className="contentInfo ng-binding"> Jordan Peele </p>
+              </div>
+              {/* Finish row Info */}
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Diễn viên</p>
+                <p className="contentInfo ng-binding">Winston Duke, Lupita Nyong`o</p>
+              </div>
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Thể Loại</p>
+                <p className="contentInfo ng-binding">Kinh dị</p>
+              </div>
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Định dạng</p>
+                <p className="contentInfo ng-binding">2D/Digital</p>
+              </div>
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Quốc Gia SX</p>
+                <p className="contentInfo ng-binding">Mỹ</p>
+              </div>
+            </div>
+            <div className="col-sm-6 col-xs-12 film right mt100">
+              <div className="row rowLeftInfo">
+                <p className="contentTitle">Nội dung</p>
+              </div>
+              {/* Start row Content */}
+              <div className="row rowLeftInfo">
+                <p className="contentInfoFull description ng-binding ">
+                  Chúng Ta là tác phẩm tiếp theo được chắp bút và cầm trịch bởi Jordan Peele - đạo diễn của Get Out. Bộ phim theo chân một gia đình da màu 4 người đang trong kỳ nghỉ hè yên bình thì chạm trán 4 bản sao đáng sợ của chính họ. Câu hỏi được đặt ra ở đây là những thực thể này từ đâu xuất hiện? Liệu chúng có liên quan gì đến quá khứ của nhà Wilson? Và họ sẽ phải làm thế nào để có thể loại bỏ được mối nguy hiểm đe dọa đến tính mạng đang đeo bám cả gia đình?</p>
+              </div>
+              {/* Start row Content */}
+              <div className="row rowLeftInfo">
+                <p className="contentInfoFull" />
+              </div>
+            </div>
+          </div>
+
+
+        </div>
+
+
+      </div>
+
+
+
+    )
+  }
+}
+
