+++ src/pages/HomePage.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { loadPackages } from '../utils';
+
 import {
   HeaderNavigation,
   Loader,
