+++ src/pages/index.js
@@ -2,5 +2,5 @@ export { default as HomePage } from './HomePage';
 export { default as NotFoundPage } from './NotFoundPage';
 export { default as MoviesPage } from './MoviesPage';
 export { default as ContactPage } from './ContactPage';
-export { default as SchedualPage } from './SchedualPage';
+export { default as MovieShowtimes } from './MovieShowtimes';
