+++ src/styles/schedualPage.css
@@ -1,35 +1,39 @@
-.detailContentCard{
+.detailContentCard {
     border: none;
-    
 }
+
 #detailMain #detailMainTop .detailMainInfo {
     position: absolute;
     top: 50%;
     width: 100%;
     left: 50%;
-    transform: translate3d(-50%,-50%,0);
+    transform: translate3d(-50%, -50%, 0);
 }
 
-.mt100{
-    margin-top:100px;
-}
-.mt150{
-    margin-top:150px;
+.mt100 {
+    margin-top: 100px;
 }
 
-.w100{
-    width:100%;
-}
-.mb{
-    margin-bottom:100px
+.mt150 {
+    margin-top: 150px;
 }
 
-.dblock{
+.w100 {
+    width: 100%;
+}
+
+.mb {
+    margin-bottom: 100px
+}
+
+.dblock {
     display: -webkit-inline-box !important;
 }
+
 .detailMainInfo1 {
     color: #e9e9e9;
 }
+
 .c100 .circleBorder {
     width: 111px;
     height: 111px;
@@ -40,6 +44,7 @@
     -webkit-border-radius: 50%;
     position: absolute;
 }
+
 .c100 {
     position: relative;
     font-size: 126px;
@@ -47,17 +52,17 @@
     height: 1em;
     border-radius: 50%;
     margin: auto;
-    background-color: rgba(0,0,0,.4);
+    background-color: rgba(0, 0, 0, .4);
 }
 
-
-.bgdetail
-{
+.bgdetail {
     height: 20%;
 }
+
 #detailCinemaComplex {
     padding-top: 40px;
 }
+
 #detailCinemaComplex .listPCinemas {
     background: #fff;
     min-height: 713px;
@@ -69,98 +74,125 @@
     border-bottom-left-radius: 4px;
     width: 28%;
 }
-.center{
 
+.center {
     text-align: center
 }
+
 #detailCinemaComplex .listPCinemas img {
     width: 50px;
     margin-right: 10px;
 }
+
 #detailMain #detailCinemaComplex .listPCinemas {
     width: 100%;
     min-height: 0;
     border-top: none;
     border-bottom: none;
 }
+
 body {
     background-color: #fff;
     font-family: 'SF Text Regular';
     overflow-x: hidden;
 }
+
 .tab-content>.active {
     display: block;
 }
+
 .tab-content>.tab-pane {
     display: none;
 }
+
 .fade.in {
     opacity: 1;
 }
+
 .fade {
     opacity: 0;
     -webkit-transition: opacity .15s linear;
     -o-transition: opacity .15s linear;
     transition: opacity .15s linear;
 }
-.ml{
-    margin-left:100px;
+
+.ml {
+    margin-left: 100px;
 }
+
 .backgroundMain {
     position: relative;
 }
+
 #detailMain .contentMain {
-    transform: translate(0,-70px);
+    transform: translate(0, -70px);
 }
+
 .mainMaxWidth2 {
     max-width: 870px;
     float: none;
     margin: auto;
     clear: both;
 }
-section {padding:2rem;}
-.bg-solid {background-color:#0a2029;color:#fff;}
-.bg-gray {background-color:#cccccc;}
 
+section {
+    padding: 2rem;
+}
+
+.bg-solid {
+    background-color: #0a2029;
+    color: #fff;
+}
+
+.bg-gray {
+    background-color: #cccccc;
+}
 
 .row {
     margin-left: 0;
     margin-right: 0;
 }
- .detailMainStyle {
+
+.detailMainStyle {
     margin-top: 40px;
 }
+
 .dark .contentInfo {
     color: #e9e9e9;
 }
 
- .contentInfo {
+.contentInfo {
     width: 40%;
     float: left;
     font-size: 15px;
     color: #fff;
 }
- .left {
+
+.left {
     margin-bottom: 0;
 }
- .film {
+
+.film {
     padding: 0 7px;
     margin-bottom: 15px;
 }
- .contentTitle {
+
+.contentTitle {
     width: 35%;
     float: left;
     font-size: 16px;
     font-family: 'SF Medium';
     color: #fff;
-  
 }
+
 #detailMain .rowLeftInfo {
     margin-bottom: 10px;
 }
+
 .contentInfoFull {
     color: #e9e9e9!important;
 }
+
 .contentTitleFull {
     width: 100%;
     float: left;
@@ -172,24 +204,23 @@ section {padding:2rem;}
     margin-bottom: 20px;
     color: #fff;
 }
-.modal-size{
 
+.modal-size {
     overflow-y: auto;
-     width: '800px';
-     height: '553px';
-     
+    width: '800px';
+    height: '553px';
 }
-.modal_design{
 
-    
-    right:0;
-    bottom:0;
-    left:0;
-    
+.modal_design {
+    right: 0;
+    bottom: 0;
+    left: 0;
 }
-.ml20{
+
+.ml20 {
     margin-left: 20px;
 }
-.mr5{
+
+.mr5 {
     margin-right: 5px;
 }
